confirm deletion

+++ dashBoard/scripts/quetinos-level-two.js
@@ -1,7 +1,7 @@
-const table = document.querySelector("table tbody");
+const table = document.querySelector('table tbody');
 let currentEditButton;
 //getting local storage level-one data
-let allQuestions = JSON.parse(localStorage.getItem("level-two"));
+let allQuestions = JSON.parse(localStorage.getItem('level-two'));
 //loading all questions
 onload = loadAllQuestios();
 function loadAllQuestios() {
@@ -10,11 +10,11 @@ function loadAllQuestios() {
   if (keys.length >= 3) {
     for (const q in allQuestions) {
       //skipping the data of "levelID" and "started" keys
-      if (q != "levleID" && q != "started") {
-        let qText = allQuestions[q]["questionContext"];
-        let qAnswer = allQuestions[q]["correctAnswers"];
+      if (q != 'levleID' && q != 'started') {
+        let qText = allQuestions[q]['questionContext'];
+        let qAnswer = allQuestions[q]['correctAnswers'];
         //creating a new table row and adding new questions to the table
-        let row = document.createElement("tr");
+        let row = document.createElement('tr');
         row.innerHTML = `
                     <td>${q}</td>
                     <td>${qText}</td>
@@ -34,15 +34,15 @@ function loadAllQuestios() {
                 `;
         table.appendChild(row);
 
-        console.log(allQuestions[q]["questionContext"]);
+        console.log(allQuestions[q]['questionContext']);
       }
     }
   }
 }
 
 function addQuestion() {
-  const questionText = document.getElementById("questionContext").value;
-  const questionAnswer = document.getElementById("questionAnswer").value;
+  const questionText = document.getElementById('questionContext').value;
+  const questionAnswer = document.getElementById('questionAnswer').value;
   if (questionText && questionAnswer) {
     createQuestions(questionText, questionAnswer, 2);
     location.reload();
@@ -52,28 +52,46 @@ function deleteQuestion(button) {
   const row = button.parentNode.parentNode;
   //getting the question id
   QID = row.cells[0];
-  `1`;
-  console.log(QID.innerText);
-  deleteLevelQuestion(2, QID.innerText);
-  row.parentNode.removeChild(row);
+  Swal.fire({
+    title: 'Are you sure?',
+    text: 'You will not be able to recover this participant!',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonText: 'Yes, delete it!',
+    cancelButtonText: 'No, keep it',
+    confirmButtonColor: '#fc0102',
+  }).then((result) => {
+    if (result.isConfirmed) {
+      deleteLevelQuestion(2, QID.innerText);
+      row.parentNode.removeChild(row);
+      Swal.fire({
+        title: 'Deleted',
+        text: 'Participant has been deleted.',
+        icon: 'success',
+        confirmButtonColor: '#3085d6',
+      });
+    } else if (result.dismiss === Swal.DismissReason.cancel) {
+      Swal.fire('Cancelled', 'Participant deletion was cancelled', 'info');
+    }
+  });
 }
 
 function openEditModal(button) {
   currentEditButton = button;
-  document.getElementById("editModal").style.display = "block";
+  document.getElementById('editModal').style.display = 'block';
   const questionText = button.parentNode.parentNode.children[1].textContent;
   const questionAnswer = button.parentNode.parentNode.children[2].textContent;
-  document.getElementById("editText").value = questionText;
-  document.getElementById("editAnswer").value = questionAnswer;
+  document.getElementById('editText').value = questionText;
+  document.getElementById('editAnswer').value = questionAnswer;
 }
 
 function closeEditModal() {
-  document.getElementById("editModal").style.display = "none";
+  document.getElementById('editModal').style.display = 'none';
 }
 
 function saveEdit() {
-  const newText = document.getElementById("editText").value;
-  const newQuestionAnswer = document.getElementById("editAnswer").value;
+  const newText = document.getElementById('editText').value;
+  const newQuestionAnswer = document.getElementById('editAnswer').value;
   if (currentEditButton && newText && newQuestionAnswer) {
     currentEditButton.parentNode.parentNode.children[1].textContent = newText;
     currentEditButton.parentNode.parentNode.children[2].textContent =
@@ -83,7 +101,7 @@ function saveEdit() {
 }
 
 window.onclick = function (event) {
-  if (event.target == document.getElementById("editModal")) {
+  if (event.target == document.getElementById('editModal')) {
     closeEditModal();
   }
 };
